+++ config/configmapprovider/default_test.go
@@ -27,7 +27,7 @@ import (
 
 func TestDefaultMapProvider(t *testing.T) {
 	mp := NewDefault("testdata/default-config.yaml", nil)
-	retr, err := mp.Retrieve(context.Background())
+	retr, err := mp.Retrieve(context.Background(), nil)
 	require.NoError(t, err)
 
 	expectedMap, err := config.NewMapFromBuffer(strings.NewReader(`
@@ -37,14 +37,16 @@ exporters:
   otlp:
     endpoint: "localhost:4317"`))
 	require.NoError(t, err)
-	assert.Equal(t, expectedMap, retr.Get())
+	m, err := retr.Get(context.Background())
+	require.NoError(t, err)
+	assert.Equal(t, expectedMap, m)
 
-	assert.NoError(t, mp.Close(context.Background()))
+	assert.NoError(t, mp.Shutdown(context.Background()))
 }
 
 func TestDefaultMapProvider_AddNewConfig(t *testing.T) {
 	mp := NewDefault("testdata/default-config.yaml", []string{"processors.batch.timeout=2s"})
-	cp, err := mp.Retrieve(context.Background())
+	cp, err := mp.Retrieve(context.Background(), nil)
 	require.NoError(t, err)
 
 	expectedMap, err := config.NewMapFromBuffer(strings.NewReader(`
@@ -55,16 +57,18 @@ exporters:
   otlp:
     endpoint: "localhost:4317"`))
 	require.NoError(t, err)
-	assert.Equal(t, expectedMap, cp.Get())
+	m, err := cp.Get(context.Background())
+	require.NoError(t, err)
+	assert.Equal(t, expectedMap, m)
 
-	assert.NoError(t, mp.Close(context.Background()))
+	assert.NoError(t, mp.Shutdown(context.Background()))
 }
 
 func TestDefaultMapProvider_OverwriteConfig(t *testing.T) {
 	mp := NewDefault(
 		"testdata/default-config.yaml",
 		[]string{"processors.batch.timeout=2s", "exporters.otlp.endpoint=localhost:1234"})
-	cp, err := mp.Retrieve(context.Background())
+	cp, err := mp.Retrieve(context.Background(), nil)
 	require.NoError(t, err)
 
 	expectedMap, err := config.NewMapFromBuffer(strings.NewReader(`
@@ -75,24 +79,26 @@ exporters:
   otlp:
     endpoint: "localhost:1234"`))
 	require.NoError(t, err)
-	assert.Equal(t, expectedMap, cp.Get())
+	m, err := cp.Get(context.Background())
+	require.NoError(t, err)
+	assert.Equal(t, expectedMap, m)
 
-	assert.NoError(t, mp.Close(context.Background()))
+	assert.NoError(t, mp.Shutdown(context.Background()))
 }
 
 func TestDefaultMapProvider_InexistentFile(t *testing.T) {
 	mp := NewDefault("testdata/otelcol-config.yaml", nil)
 	require.NotNil(t, mp)
-	_, err := mp.Retrieve(context.Background())
+	_, err := mp.Retrieve(context.Background(), nil)
 	require.Error(t, err)
 
-	assert.NoError(t, mp.Close(context.Background()))
+	assert.NoError(t, mp.Shutdown(context.Background()))
 }
 
 func TestDefaultMapProvider_EmptyFileName(t *testing.T) {
 	mp := NewDefault("", nil)
-	_, err := mp.Retrieve(context.Background())
+	_, err := mp.Retrieve(context.Background(), nil)
 	require.Error(t, err)
 
-	assert.NoError(t, mp.Close(context.Background()))
+	assert.NoError(t, mp.Shutdown(context.Background()))
 }