+++ config/configmapprovider/expand_test.go
@@ -57,11 +57,13 @@ func TestExpand(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			// Retrieve the config
 			emp := NewExpand(NewFile(path.Join("testdata", test.name)))
-			cp, err := emp.Retrieve(context.Background())
+			cp, err := emp.Retrieve(context.Background(), nil)
 			require.NoError(t, err, "Unable to get config")
 
 			// Test that expanded configs are the same with the simple config with no env vars.
-			assert.Equal(t, expectedCfgMap.ToStringMap(), cp.Get().ToStringMap())
+			m, err := cp.Get(context.Background())
+			require.NoError(t, err)
+			assert.Equal(t, expectedCfgMap.ToStringMap(), m.ToStringMap())
 		})
 	}
 }
@@ -75,7 +77,7 @@ func TestExpand_EscapedEnvVars(t *testing.T) {
 
 	// Retrieve the config
 	emp := NewExpand(NewFile(path.Join("testdata", "expand-escaped-env.yaml")))
-	cp, err := emp.Retrieve(context.Background())
+	cp, err := emp.Retrieve(context.Background(), nil)
 	require.NoError(t, err, "Unable to get config")
 
 	expectedMap := map[string]interface{}{
@@ -95,5 +97,7 @@ func TestExpand_EscapedEnvVars(t *testing.T) {
 			// escaped $ alone
 			"recv.7": "$",
 		}}
-	assert.Equal(t, expectedMap, cp.Get().ToStringMap())
+	m, err := cp.Get(context.Background())
+	require.NoError(t, err)
+	assert.Equal(t, expectedMap, m.ToStringMap())
 }