+++ config/configmapprovider/provider.go
@@ -20,42 +20,78 @@ import (
 	"go.opentelemetry.io/collector/config"
 )
 
-// Provider is an interface that helps providing configuration's parser.
-// Implementations may load the parser from a file, a database or any other source.
+// Provider is an interface that helps to retrieve a config map and watch for any
+// changes to the config map. Implementations may load the config from a file,
+// a database or any other source.
 type Provider interface {
 	// Retrieve goes to the configuration source and retrieves the selected data which
 	// contains the value to be injected in the configuration and the corresponding watcher that
 	// will be used to monitor for updates of the retrieved value.
-	Retrieve(ctx context.Context) (Retrieved, error)
+	//
+	// onChange callback is called when the config changes. onChange may be called from
+	// a different go routine. After onChange is called Retrieved.Get should be called
+	// to get the new config. See description of Retrieved for more details.
+	// onChange may be nil, which indicates that the caller is not interested in
+	// knowing about the changes.
+	//
+	// If ctx is cancelled should return immediately with an error.
+	// Should never be called concurrently with itself or with Shutdown.
+	Retrieve(ctx context.Context, onChange func(*ChangeEvent)) (Retrieved, error)
 
-	// Close signals that the configuration for which it was used to retrieve values is no longer in use
-	// and the object should close and release any watchers that it may have created.
-	// This method must be called when the service ends, either in case of success or error.
-	Close(ctx context.Context) error
+	// Shutdown signals that the configuration for which this Provider was used to
+	// retrieve values is no longer in use and the Provider should close and release
+	// any resources that it may have created.
+	//
+	// This method must be called when the Collector service ends, either in case of
+	// success or error. Retrieve cannot be called after Shutdown.
+	//
+	// Should never be called concurrently with itself or with Retrieve.
+	// If ctx is cancelled should return immediately with an error.
+	Shutdown(ctx context.Context) error
 }
 
-// Retrieved holds the result of a call to the Retrieve method of a Session object.
+// Retrieved holds the result of a call to the Retrieve method of a Provider object.
+//
+// The typical usage is the following:
+//
+//		r := mapProvider.Retrieve()
+//		r.Get()
+//		// wait for onChange() to be called.
+//		r.Close()
+//		r = mapProvider.Retrieve()
+//		r.Get()
+//		// wait for onChange() to be called.
+//		r.Close()
+//		// repeat Retrieve/Get/wait/Close cycle until it is time to shut down the Collector process.
+//		// ...
+//		mapProvider.Shutdown()
 type Retrieved interface {
-	// Get returns the Map.
-	Get() *config.Map
+	// Get returns the config Map.
+	// If Close is called before Get or concurrently with Get then Get
+	// should return immediately with ErrSessionClosed error.
+	// Should never be called concurrently with itself.
+	// If ctx is cancelled should return immediately with an error.
+	Get(ctx context.Context) (*config.Map, error)
+
+	// Close signals that the configuration for which it was used to retrieve values is
+	// no longer in use and the object should close and release any watchers that it
+	// may have created.
+	//
+	// This method must be called when the service ends, either in case of success or error.
+	//
+	// Should never be called concurrently with itself.
+	// May be called before, after or concurrently with Get.
+	// If ctx is cancelled should return immediately with an error.
+	//
+	// Calling Close on an already closed object should have no effect and should return nil.
+	Close(ctx context.Context) error
 }
 
-// WatchableRetrieved is an extension for Retrieved that if implemented,
-// the Retrieved value supports monitoring for updates.
-type WatchableRetrieved interface {
-	Retrieved
-
-	// WatchForUpdate waits for updates on any of the values retrieved from config sources.
-	// It blocks until configuration updates are received and can
-	// return an error if anything fails. WatchForUpdate is used once during the
-	// first evaluation of the configuration and is not used to watch configuration
-	// changes continuously.
-	WatchForUpdate() error
-
-	// Close signals that the configuration for which it was used to retrieve values is no longer in use
-	// and the object should close and release any watchers that it may have created.
-	// This method must be called when the service ends, either in case of success or error.
-	// The method may be called while WatchForUpdate() class is in progress and is blocked.
-	// In that case WatchForUpdate() method must abort as soon as possible and return ErrSessionClosed.
-	Close(ctx context.Context) error
+// ChangeEvent describes the particular change event that happened with the config.
+// TODO: see if this can be eliminated.
+type ChangeEvent struct {
+	// Error is nil if the config is changed and needs to be re-fetched using Get.
+	// It is set to configsource.ErrSessionClosed if Close was called.
+	// Any other error indicates that there was a problem with retrieving the config.
+	Error error
 }