+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## 🛑 Breaking changes 🛑
 
+- Remove `pdata.AttributeMap.InitFromMap` (#4429)
+
+## v0.39.0 Beta
+
+## 🛑 Breaking changes 🛑
+
 - Remove deprecated config (already no-op) `ballast_size_mib` in memorylimiterprocessor (#4365)
 - Remove `config.Receivers`, `config.Exporters`, `config.Processors`, and `config.Extensions`. Use map directly (#4344)
 - Remove `component.BaseProcessorFactory`, use `processorhelper.NewFactory` instead (#4175)
@@ -27,6 +33,7 @@
 
 ## 💡 Enhancements 💡
 - Supports more compression methods(`snappy` and `zstd`) for configgrpc, in addition to current `gzip` (#4088)
+- Moved the OpenTelemetry Collector Builder to core (#4307)
 
 ## 🧰 Bug fixes 🧰